+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# v1.1.0
+## 2026-01-07
+- fix: correct handling of state/sensor updates, so the states of the accessory in the Home app refreshes correctly
+- fix: correct mutex usage
+- fix: supports Node.js v24 (fixes #26)
+- chore: outdated dependencies
+
 # v1.0.11
 ## 2024-07-21
 - fix: added Canada to the custom UI